extract logic out of stock_location model into the PACKER
