+++ core/spec/models/spree/stock/packer_spec.rb
@@ -1,0 +1,70 @@
+require 'spec_helper'
+
+module Spree
+  module Stock
+    describe Packer do
+      let(:order) { create(:order_with_line_items, line_items_count: 5) }
+      let(:stock_location) { create(:stock_location) }
+
+      subject { Packer.new(stock_location, order) }
+
+      before :all do
+        order.reload
+        order.line_items.each do |line_item|
+          create(:stock_item,
+                 variant: line_item.variant,
+                 count_on_hand: 5,
+                 stock_location: stock_location)
+        end
+      end
+
+      context 'default_package' do
+        it 'contains all the items' do
+          package = subject.default_package
+          package.contents.size.should eq 5
+          package.weight.should > 0
+        end
+
+        it 'variants are added as backordered without enough on_hand' do
+          subject.should_receive(:stock_status).exactly(5).times.and_return([2,3])
+          package = subject.default_package
+          package.on_hand.size.should eq 5
+          package.backordered.size.should eq 5
+        end
+      end
+
+      xit 'builds an array of packages' do
+        order.reload #temp fix for bug where line_items weren't loaded
+        packages = subject.packages(order)
+        packages.size.should eq 1
+        packages.first.contents.size.should eq 5
+      end
+
+      context 'stock_status' do
+        before do
+          @stock_item = build(:stock_item, count_on_hand: 10)
+          subject.stock_location.should_receive(:stock_item).and_return(@stock_item)
+        end
+
+        it 'all on_hand with no backordered' do
+          on_hand, backordered = subject.send(:stock_status, double, 5)
+          on_hand.should eq 5
+          backordered.should eq 0
+        end
+
+        it 'some on_hand with some backordered' do
+          on_hand, backordered = subject.send(:stock_status, double, 20)
+          on_hand.should eq 10
+          backordered.should eq 10
+        end
+
+        it 'zero on_hand with all backordered' do
+          @stock_item.stub(count_on_hand: 0)
+          on_hand, backordered = subject.send(:stock_status, double, 20)
+          on_hand.should eq 0
+          backordered.should eq 20
+        end
+      end
+    end
+  end
+end
